requirements.txt
Django==4.1.1
Django==4.2.4
django_crispy_forms==2.0
django_recaptcha==3.0.0
